requirements.txt
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
coverage==5.5
deprecation==2.1.0
et-xmlfile==1.1.0
eyed3==0.9.7
ffmpeg==1.4
filetype==1.2.0
idna==3.4
mutagen==1.46.0
numpy==1.23.5
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pycodestyle==2.10.0
pydub==0.25.1
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.6
requests==2.28.2
six==1.16.0
soupsieve==2.4
toml==0.10.2
urllib3==1.26.14
